+++ src/test/java/com/bigfish/securitydemonstration/SecurityDemonstrationApplicationTests.java
@@ -1,4 +1,4 @@
-package com.bigfish.securitydemonstrationgezaczimeth;
+package com.bigfish.securitydemonstration;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
